Moved corsFilter to SecurityConfig
